- hostmask.c, rng_mt.c: avoid use of long int types
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid-8@1343 82007160-df01-0410-b94d-b575c5fd34c7

+++ include/hostmask.h
@@ -35,7 +35,7 @@ enum
 struct HostMaskEntry
 {
   int type, subtype;
-  unsigned long precedence;
+  unsigned int precedence;
   char *hostmask;
   void *data;
   struct HostMaskEntry *next, *nexthash;
@@ -93,7 +93,7 @@ struct AddressRec
   int type;
 
   /* Higher precedences overrule lower ones... */
-  unsigned long precedence;
+  unsigned int precedence;
 
   /* Only checked if !(type & 1)... */
   const char *username;
